IRON

Struct static::Cache

pub struct Cache {
    pub duration: Duration,
}

A modifier for Static to specify a response's cache-control.

Fields

duration

The length of time the file should be cached for.

Methods

impl Cache

fn new(duration: Duration) -> Cache

Create a new instance of Cache with a given duration.